后台页面测试增删用户

+++ springvue-backend/src/main/java/com/bx/springvue/Service/Implement/UserServiceImplement.java
@@ -34,8 +34,13 @@ public class UserServiceImplement implements UserService {
     }
 
     @Override
-    public void deleteUserbyname(String name) {
-        dataMapper.deleteUserbyname(name);
+    public boolean deleteUserbyname(String name) {
+        if(name==null){
+            return false;
+        }else{
+            dataMapper.deleteUserbyname(name);
+            return true;
+        }
     }
 
     @Override
